Reorder homepage + accessibility tweaks (#2528)
* Reorder homepage + accessibility tweaks

* Homepage refactor

* Pot, hide mobile banny

* Fix text

* Ux review feedback

* Flip CTA

* add href links to homepage components that are missing them
