+++ src/pages/home/Faq.tsx
@@ -1,14 +1,12 @@
 import { Collapse } from 'antd'
-import CollapsePanel from 'antd/lib/collapse/CollapsePanel'
 import Image from 'next/image'
-
 import QAs from './QAs'
 
 export default function Faq() {
   return (
     <Collapse defaultActiveKey={QAs.length ? 0 : undefined} accordion>
       {QAs().map((qa, i) => (
-        <CollapsePanel header={qa.q} key={i}>
+        <Collapse.Panel header={qa.q} key={i}>
           {qa.a && <div>{qa.a}</div>}
           {qa.img && (
             <Image
@@ -19,7 +17,7 @@ export default function Faq() {
               loading="lazy"
             />
           )}
-        </CollapsePanel>
+        </Collapse.Panel>
       ))}
     </Collapse>
   )
