+++ src/pages/privacy.page.tsx
@@ -1,6 +1,7 @@
 import { AppWrapper } from 'components/common'
 
 import ExternalLink from 'components/ExternalLink'
+import { LAYOUT_MAX_WIDTH_PX } from 'constants/styles/layouts'
 
 export default function PrivacyPolicyPage() {
   return (
@@ -12,7 +13,13 @@ export default function PrivacyPolicyPage() {
 
 function PrivacyPolicy() {
   return (
-    <div style={{ padding: '40px', maxWidth: '1080px', margin: '0 auto' }}>
+    <div
+      style={{
+        padding: '40px',
+        maxWidth: LAYOUT_MAX_WIDTH_PX,
+        margin: '0 auto',
+      }}
+    >
       <h1>Privacy Policy</h1>
       <p>
         We want to process as little personal information as possible when you
