+++ src/styles/antd-overrides/collapse.scss
@@ -7,7 +7,7 @@
 
     .ant-collapse-header {
       font-weight: 500;
-      color: var(--text-header);
+      color: var(--text-primary);
     }
   }
